+++ src/TaskDecoder.ts
@@ -281,15 +281,15 @@ export const array = <A>(item: TaskDecoder<unknown, A>): TaskDecoder<unknown, Ar
  * @category combinators
  * @since 2.2.7
  */
-export const krecord = <I, A>(codomain: TaskDecoder<I, A>): TaskDecoder<Record<string, I>, Record<string, A>> =>
-  K.record(M)((k, e) => FS.of(DE.key(k, DE.optional, e)))(codomain)
+export const values = <I, A>(codomain: TaskDecoder<I, A>): TaskDecoder<Record<string, I>, Record<string, A>> =>
+  K.values(M)((k, e) => FS.of(DE.key(k, DE.optional, e)))(codomain)
 
 /**
  * @category combinators
  * @since 2.2.8
  */
 export const record = <A>(codomain: TaskDecoder<unknown, A>): TaskDecoder<unknown, Record<string, A>> =>
-  pipe(UnknownRecord, compose(krecord(codomain)))
+  pipe(UnknownRecord, compose(values(codomain)))
 
 /**
  * @category combinators